app/requirements.txt
yt-dlp>=2022.6.29
flask
mysql-connector-python